refactor: address PR review feedback for feed diversity mechanisms
PR Review Fixes:
- S1: Reverted unrelated formatting change in MarketsTradingTerminal.tsx
- M1: Extracted diversity utilities to shared feed-diversity.ts module
  - Moved shuffleWithNoConsecutive, createDiscourseActionDeck, ActionDiversityTracker
  - Tests now import from shared module instead of duplicating code
  - Exported from utils/index.ts for reuse
- S2: Made maxRecent configurable via NPC_MAX_RECENT_ACTIONS env var
  - Added maxRecentActions to NPC_DIVERSITY_CONFIG
  - Updated .env.example with documentation
- S3: Fixed timestamp naming confusion
  - Removed redundant 'const now = baseNow' assignment
  - Consistently use 'baseNow' throughout npc-social-engagement-service.ts
- M2: Documented engagement volume trade-off
  - Added NOTE in ActionDiversityTracker class explaining skip behavior

All tests pass (27/27), TypeScript compiles without errors.
