+++ .env.example
@@ -455,6 +455,7 @@ NPC_TICK_BATCH_SIZE=20
 # NPC Feed Diversity Configuration (TikTok-inspired)
 # NPC_MAX_CONSECUTIVE_SAME_ACTION - Max consecutive same action type, e.g. likes (default: 1)
 # NPC_TIMESTAMP_STAGGER_MS - Stagger window in ms for organic pacing, 0 to disable (default: 300000)
+# NPC_MAX_RECENT_ACTIONS - Number of recent actions to track for diversity (default: 5)
 
 # Maximum consecutive errors before stopping NPC tick
 NPC_TICK_MAX_ERRORS=5
