+++ packages/engine/src/__tests__/unit/feed-diversity-mechanisms.test.ts
@@ -15,143 +15,11 @@
  */
 
 import { describe, expect, test } from 'bun:test';
-
-// =============================================================================
-// SHUFFLE WITH NO CONSECUTIVE - Import the function for testing
-// =============================================================================
-
-/**
- * Fisher-Yates shuffle with constraint: no consecutive same values.
- * This is a copy of the function from post-generation-helpers.ts for testing.
- *
- * Includes dead-end prevention: before selecting a candidate, we verify that
- * removing it won't make the remaining items impossible to arrange without
- * consecutive duplicates (maxCount <= ceil(length / 2)).
- */
-function shuffleWithNoConsecutive<T>(arr: T[], random: () => number): T[] {
-  if (arr.length <= 1) return [...arr];
-
-  const result: T[] = [];
-  const remaining = [...arr];
-
-  // Track counts of each value for efficient dead-end detection
-  const countByValue = new Map<T, number>();
-  for (const item of remaining) {
-    countByValue.set(item, (countByValue.get(item) ?? 0) + 1);
-  }
-
-  while (remaining.length > 0) {
-    const lastItem = result[result.length - 1];
-
-    // Find valid candidates (different from last item, or any if first pick)
-    const validIndices: number[] = [];
-    for (let i = 0; i < remaining.length; i++) {
-      if (result.length === 0 || remaining[i] !== lastItem) {
-        validIndices.push(i);
-      }
-    }
-
-    // Filter to safe candidates: those whose removal keeps ordering possible
-    const safeCandidates = validIndices.filter((idx) => {
-      const value = remaining[idx];
-      const newLength = remaining.length - 1;
-      if (newLength === 0) return true;
-
-      let maxCountAfter = 0;
-      for (const [v, count] of countByValue) {
-        const newCount = v === value ? count - 1 : count;
-        if (newCount > maxCountAfter) maxCountAfter = newCount;
-      }
-
-      return maxCountAfter <= Math.ceil(newLength / 2);
-    });
-
-    // Use safe candidates if any, otherwise fall back to valid, then any
-    const candidates =
-      safeCandidates.length > 0
-        ? safeCandidates
-        : validIndices.length > 0
-          ? validIndices
-          : remaining.map((_, idx) => idx);
-
-    // Pick random from candidates
-    const pickIdx = candidates[Math.floor(random() * candidates.length)]!;
-    const pickedValue = remaining[pickIdx]!;
-    result.push(pickedValue);
-
-    // Update counts and remaining array
-    countByValue.set(pickedValue, (countByValue.get(pickedValue) ?? 1) - 1);
-    remaining.splice(pickIdx, 1);
-  }
-
-  return result;
-}
-
-/**
- * Creates a stratified action deck with guaranteed ratios.
- */
-function createDiscourseActionDeck(
-  totalSlots: number,
-  quoteRatio: number,
-  random: () => number
-): Array<'quote' | 'reply'> {
-  if (totalSlots <= 0) return [];
-
-  // Clamp quoteRatio to valid [0, 1] range
-  const clampedRatio = Math.max(0, Math.min(1, quoteRatio));
-  const quoteCount = Math.round(totalSlots * clampedRatio);
-  const replyCount = totalSlots - quoteCount;
-
-  const deck: Array<'quote' | 'reply'> = [
-    ...Array(quoteCount).fill('quote' as const),
-    ...Array(replyCount).fill('reply' as const),
-  ];
-
-  return shuffleWithNoConsecutive(deck, random);
-}
-
-// =============================================================================
-// ACTION DIVERSITY TRACKER - Copy for testing
-// =============================================================================
-
-type EngagementActionType = 'like' | 'share' | 'comment';
-
-class ActionDiversityTracker {
-  private recentActions: EngagementActionType[] = [];
-  private readonly maxRecent: number;
-  private readonly maxConsecutive: number;
-
-  constructor(maxRecent = 5, maxConsecutive = 2) {
-    this.maxRecent = Math.max(maxRecent, maxConsecutive);
-    this.maxConsecutive = Math.max(1, maxConsecutive);
-  }
-
-  recordAction(type: EngagementActionType): void {
-    this.recentActions.push(type);
-    if (this.recentActions.length > this.maxRecent) {
-      this.recentActions.shift();
-    }
-  }
-
-  shouldSkipForDiversity(type: EngagementActionType): boolean {
-    const lastN = this.recentActions.slice(-this.maxConsecutive);
-    return (
-      lastN.length >= this.maxConsecutive && lastN.every((t) => t === type)
-    );
-  }
-
-  getDistribution(): Record<EngagementActionType, number> {
-    const dist: Record<EngagementActionType, number> = {
-      like: 0,
-      share: 0,
-      comment: 0,
-    };
-    for (const action of this.recentActions) {
-      dist[action]++;
-    }
-    return dist;
-  }
-}
+import {
+  ActionDiversityTracker,
+  createDiscourseActionDeck,
+  shuffleWithNoConsecutive,
+} from '../../utils/feed-diversity';
 
 // =============================================================================
 // TESTS: shuffleWithNoConsecutive
